+++ src/QuerySerializer/Rfc3986Serializer.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace GuzzleHttp\Command\Guzzle\QuerySerializer;
 
+use GuzzleHttp\Command\Guzzle\NonFiniteFloats;
+
 /**
  * @final
  */
@@ -21,7 +23,7 @@ class Rfc3986Serializer implements QuerySerializerInterface
      */
     public function aggregate(array $queryParams): string
     {
-        $queryString = http_build_query($queryParams, '', '&', PHP_QUERY_RFC3986);
+        $queryString = http_build_query(NonFiniteFloats::normalizeAll($queryParams, 'a query location value'), '', '&', PHP_QUERY_RFC3986);
 
         if ($this->removeNumericIndices) {
             $queryString = preg_replace('/%5B[0-9]+%5D/simU', '%5B%5D', $queryString);